chore(Tests): Cleanup test code
