+++ src/tests/components/withKeycloak.spec.js
@@ -1,31 +1,13 @@
-import React, { Component } from 'react'
+import React from 'react'
 import * as rtl from '@testing-library/react'
 import '@testing-library/jest-dom/extend-expect'
 
-import KeycloakProvider from '../../lib/components/Provider'
-import { withKeycloak } from '../../lib'
+import { KeycloakProvider, withKeycloak } from '../../lib'
 
-const createKeycloakStub = () => ({
-  init: () => {},
-  updateToken: () => {}
-})
-
-const createChild = (testedProps = []) => {
-  class Child extends Component {
-    render () {
-      return (
-        <div data-testid='keycloak'>
-          {Object.keys(this.props).filter(prop => testedProps.includes(prop))}
-        </div>
-      )
-    }
-  }
-
-  return Child
-}
+import { createChild, createKeycloakStub } from '../utils'
 
 describe('withKeycloak HOC', () => {
-  afterEach(() => rtl.cleanup())
+  afterEach(rtl.cleanup)
 
   it('should pass Keycloak to component props', () => {
     const keycloakStub = createKeycloakStub()
